First bits of Wizard, fix macOS build
